Move user ip to IP
