+++ Sources/QueryString.php
@@ -130,24 +130,6 @@ class QueryString
 		'xmlhttp'				=> Actions\XmlHttp::class,
 	];
 
-	/****************************
-	 * Internal static properties
-	 ****************************/
-
-	/**
-	 * IP of the current user. Typically filled with $_SERVER['REMOTE_ADDR']
-	 * or a IP supplied by a reverse proxy.
-	 * @var IP|string|null
-	 */
-	private static ?string $user_ip = null;
-
-	/**
-	 * The alternative IP of a user. Typically filled with $_SERVER['REMOTE_ADDR']
-	 * Used to check the IP in cases where a reverse proxy supplied us the main ip.
-	 * @var IP|string|null
-	 */
-	private static ?string $user_ip_alternative = null;
-
 	/***********************
 	 * Public static methods
 	 ***********************/
@@ -354,11 +336,11 @@ class QueryString
 		}
 
 		// Try to calculate their most likely IP for those people behind proxies (And the like).
-		self::setUserIPAlternative();
+		IP::setUserIPAlternative();
 
 		if (!empty(Config::$backward_compatibility)) {
-			$_SERVER['BAN_CHECK_IP'] = self::getUserIPAlternative();
-			$_SERVER['REMOTE_ADDR'] = self::getUserIP();
+			$_SERVER['BAN_CHECK_IP'] = IP::getUserIPAlternative();
+			$_SERVER['REMOTE_ADDR'] = IP::getUserIP();
 			$_SERVER['is_cli'] = Sapi::isCLI();
 		}
 
@@ -382,128 +364,6 @@ class QueryString
 
 		// Make sure HTTP_USER_AGENT is set.
 		$_SERVER['HTTP_USER_AGENT'] = isset($_SERVER['HTTP_USER_AGENT']) ? Utils::htmlspecialchars(Db::$db->unescape_string($_SERVER['HTTP_USER_AGENT']), ENT_QUOTES) : '';
-	}
-
-	/**
-	 * Determines the IP of the current user.
-	 *
-	 * What it does:
-	 * - Checks REMOTE_ADDR exists
-	 * - If proxy detection is enabled, parses headers to find a valid header.
-	 * - If proxy IP filtering is setup, validates IP is in acceptable ranges
-	 * 		- Single IP or CIDR notation allowed.
-	 * - If a valid match is found return.
-	 */
-	public static function getUserIP(): string
-	{
-		// We already figured it out once.
-		if (static::$user_ip !== null) {
-			return static::$user_ip;
-		}
-
-		if (empty($_SERVER['REMOTE_ADDR']) || $_SERVER['REMOTE_ADDR'] == 'unknown') {
-			return '';
-		}
-
-		// If we haven't specified how to handle Reverse Proxy IP headers, lets do what we always used to do.
-		if (!isset(Config::$modSettings['proxy_ip_header'])) {
-			Config::$modSettings['proxy_ip_header'] = 'autodetect';
-		}
-
-		// Which headers are we going to check for Reverse Proxy IP headers?
-		if (Config::$modSettings['proxy_ip_header'] == 'disabled') {
-			$reverseIPheaders = [];
-		} elseif (Config::$modSettings['proxy_ip_header'] == 'autodetect') {
-			$reverseIPheaders = ['HTTP_X_FORWARDED_FOR', 'HTTP_CLIENT_IP', 'HTTP_X_REAL_IP', 'HTTP_CF_CONNECTING_IP'];
-		} else {
-			$reverseIPheaders = [Config::$modSettings['proxy_ip_header']];
-		}
-
-		// Find the user's IP address. (but don't let it give you 'unknown'!)
-		foreach ($reverseIPheaders as $proxyIPheader) {
-			// Ignore if this is not set.
-			if (!isset($_SERVER[$proxyIPheader])) {
-				continue;
-			}
-
-			if (!empty(Config::$modSettings['proxy_ip_servers'])) {
-				$valid_sender = false;
-
-				foreach (explode(',', Config::$modSettings['proxy_ip_servers']) as $proxy) {
-					if (
-						$proxy == $_SERVER['REMOTE_ADDR']
-						|| (new IP($_SERVER['REMOTE_ADDR']))->matchToCIDR($proxy)
-					) {
-						$valid_sender = true;
-						break;
-					}
-				}
-
-				if (!$valid_sender) {
-					continue;
-				}
-			}
-
-			// If there are commas, get the last one.. probably.
-			if (str_contains($_SERVER[$proxyIPheader], ',')) {
-				$ips = array_reverse(explode(', ', $_SERVER[$proxyIPheader]));
-
-				// Go through each IP...
-				foreach ($ips as $i => $ip) {
-					// Make sure it's in a valid range...
-					if (preg_match('~^((0|10|172\.(1[6-9]|2[0-9]|3[01])|192\.168|255|127)\.|unknown|::1|fe80::|fc00::)~', $ip) != 0 && preg_match('~^((0|10|172\.(1[6-9]|2[0-9]|3[01])|192\.168|255|127)\.|unknown|::1|fe80::|fc00::)~', $_SERVER['REMOTE_ADDR']) == 0) {
-						if (!IP::create($_SERVER[$proxyIPheader])->isValid(FILTER_FLAG_IPV6) || preg_match('~::ffff:\d+\.\d+\.\d+\.\d+~', $_SERVER[$proxyIPheader]) !== 0) {
-							$_SERVER[$proxyIPheader] = preg_replace('~^::ffff:(\d+\.\d+\.\d+\.\d+)~', '$1', $_SERVER[$proxyIPheader]);
-
-							// Just incase we have a legacy IPv4 address.
-							// @ TODO: Convert to IPv6.
-							if (preg_match('~^((([1]?\d)?\d|2[0-4]\d|25[0-5])\.){3}(([1]?\d)?\d|2[0-4]\d|25[0-5])$~', $_SERVER[$proxyIPheader]) === 0) {
-								continue;
-							}
-						}
-
-						continue;
-					}
-
-					// Otherwise, we've got an IP!
-					return static::$user_ip = trim($ip);
-				}
-			}
-			// Otherwise just use the only one.
-			elseif (preg_match('~^((0|10|172\.(1[6-9]|2[0-9]|3[01])|192\.168|255|127)\.|unknown|::1|fe80::|fc00::)~', $_SERVER[$proxyIPheader]) == 0 || preg_match('~^((0|10|172\.(1[6-9]|2[0-9]|3[01])|192\.168|255|127)\.|unknown|::1|fe80::|fc00::)~', $_SERVER['REMOTE_ADDR']) != 0) {
-				return $_SERVER[$proxyIPheader];
-			} elseif (!IP::create($_SERVER[$proxyIPheader])->isValid(FILTER_FLAG_IPV6) || preg_match('~::ffff:\d+\.\d+\.\d+\.\d+~', $_SERVER[$proxyIPheader]) !== 0) {
-				$_SERVER[$proxyIPheader] = preg_replace('~^::ffff:(\d+\.\d+\.\d+\.\d+)~', '$1', $_SERVER[$proxyIPheader]);
-
-				// Just incase we have a legacy IPv4 address.
-				// @ TODO: Convert to IPv6.
-				if (preg_match('~^(((1?\d)?\d|2[0-4]\d|25[0-5])\.){3}(([1]?\d)?\d|2[0-4]\d|25[0-5])$~', $_SERVER[$proxyIPheader]) === 0) {
-					continue;
-				}
-			}
-		}
-
-		return static::$user_ip = $_SERVER['REMOTE_ADDR'];
-	}
-
-	/**
-	 * Determines the Alternative IP of the current user.
-	 *
-	 * Returns BAN_CHECK_IP if set, otherwise REMOTE_ADDR.
-	 */
-	public static function getUserIPAlternative(): string
-	{
-		return static::$user_ip_alternative ?? $_SERVER['REMOTE_ADDR'];
-	}
-
-	/**
-	 * Set's the alternative IP for the current user.
-	 * If nothing is provided, we use $_SERVER['REMOTE_ADDR']
-	 * @param ?string $ip
-	 */
-	public static function setUserIPAlternative(?string $ip = null): void
-	{
-		static::$user_ip_alternative = $ip ?? $_SERVER['REMOTE_ADDR'];
 	}
 
 	/**
